% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.Aspaces.R
\name{as.data.frame.ASpaces}
\alias{as.data.frame.ASpaces}
\title{Convert ASPaces to Data Frame}
\usage{
as.data.frame.ASpaces(ASpaces, remove.notrip = TRUE)
}
\arguments{
\item{ASpaces}{an object of class \code{ASpaces}}

\item{remove.notrip}{logical. See details.}
}
\value{
An object of class data.frame
}
\description{
Coerces the Input Object of Class ASpaces to a Data Frame Object
}
\details{
This conversion function converts the ASpaces class composed of a 
relatively complex structure into a simple data.frame. The existing ASpaces 
class is divided into info and trips inside the data and is stored in a list 
form for each individual, so it is easy to distinguish between data but has 
a limitation that it is difficult to see at a glance.

This function converts to data.frame to make it easier to check. 
In this process, the conversion is performed in a way that the information 
of the corresponding info for each trip is connected equally.

Among the arguments, remove.notrip determines whether to keep the trip as 
it is or to delete it for an individual that does not exist. If set to TRUE, 
which is the default value, the relevant individuals are deleted, and if set 
to FALSE, the trip data of NA is attached to the information of personal info.
}
\examples{
#load data
data(slicedata)

#convert ASpaces into data.frame
df.slicedata <- as.data.frame.ASpaces(slicedata)

#compare with original data and result
slicedata@data[[1]]
df.slicedata[1:2, ]
}
\author{
Changlock Choi (hihi7100@khu.ac.kr), Seong-Yun Hong (syhong@khu.ac.kr)
}
