% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.Tracks.R
\name{as.Tracks}
\alias{as.Tracks}
\title{Convert Objects to Class Tracks}
\usage{
as.Tracks(x, sp, varname)
}
\arguments{
\item{x}{an object of class ASpace}

\item{sp}{an object of class sp used in x}

\item{varname}{name of the column in sp's data.frame which can be matched}
}
\value{
An object of class Tracks
}
\description{
Coerces the Input Object of Class ASpaces to a Tracks Object
}
\details{
Function \code{as.Track} accepts converts \code{ASpace} objects 
to an onject of class \code{Tracks}. When converting a single \code{ASpace} 
object, unlike the \code{ASpaces} class, there is basically no sp given, so 
it is essential to enter the sp argument. 

Also, since the input of the time is fixed to the \code{POSIXct} or 
\code{POSIXt} class in the \code{Tracks} class, attention is needed that the 
existing time format is changed to the type of the corresponding class.
}
\examples{
## Not run:
# creates sample info data
testinfo <- list()

# creates sample trip data
testtrip <- data.frame(tr_id = 1:10, tr_seq = rep(NA, 10), purpose = rep(NA, 10),
                      mode = rep(NA, 10), o_type = rep(NA, 10), 
                      o_time = c(200,300,400,500,600,700,800,900,1000,1100),
                      o_zone = c(1,2,3,4,5,6,5,4,3,2),
                      d_type = rep(NA, 10),
                      d_time = c(300,400,500,600,700,800,900,1000,1100,1200),
                      d_zone = c(2,3,4,5,6,5,4,3,2,1))
                    
#constructs an object of class 'ASpace'
testASpace <- new("ASpace", info = list(id = 1), trip = testtrip)

testTracks <- as.tracksCollection(testASpace, varname = "adm_dr_cd")
}
\seealso{
\code{\link{as.Tracks}}
}
\author{
Seong-Yun Hong (syhong@khu.ac.kr)
}
