% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.tracksCollection.R
\name{as.tracksCollection}
\alias{as.tracksCollection}
\title{Convert Objects to Class tracksCollection}
\usage{
as.tracksCollection(x, sp, varname)
}
\arguments{
\item{x}{an object of class ASpaces}

\item{sp}{an object of class sp used in x}

\item{varname}{name of the column in sp's data.frame which can be matched 
with o_zone and d_zone in ASpaces}
}
\value{
An object of class tracksCollection
}
\description{
Coerces the Input Object of Class ASpaces to a tracksCollection Object
}
\details{
Function \code{as.trackCollection} accepts converts ASpaces objects 
to an onject of class \code{tracksCollection}. In the conversion since the 
individual's travel position matches the position on the spatial data, care 
must be taken in selecting 'sp'. If 'sp' is not given, the 'sp' contained in 
\code{ASpaces} is selected.
}
\examples{
## Not run:
# creates sample info data
testinfo <- list()

# creates sample trip data
testtrip <- data.frame(tr_id = 1:10, tr_seq = rep(NA, 10), purpose = rep(NA, 10),
                      mode = rep(NA, 10), o_type = rep(NA, 10), 
                      o_time = c(200,300,400,500,600,700,800,900,1000,1100),
                      o_zone = c(1,2,3,4,5,6,5,4,3,2),
                      d_type = rep(NA, 10),
                      d_time = c(300,400,500,600,700,800,900,1000,1100,1200),
                      d_zone = c(2,3,4,5,6,5,4,3,2,1))
                    
#constructs an object of class 'ASpace'
testASpace <- new("ASpace", info = list(id = 1), trip = testtrip)

#constructs an object of class 'ASpaces'
testtASpaces <- new("ASpaces", data = list(testaspace), sp = testshp)

testTC <- as.tracksCollection(testASpaces, varname = "adm_dr_cd")
}
\seealso{
\code{\link{as.Tracks}}
}
\author{
Seong-Yun Hong (syhong@khu.ac.kr)
}
