% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASpace-class.R
\docType{Class}
\name{ASpace-class}
\alias{ASpace-class}
\title{Class ASpace}
\description{
A S4 class for storing activity-space data; Contains information of one individual person
}
\details{
The slot 'info' contains demographic and socioeconomic variables
of the respondent as a list object. The list object must include
the following named element(s):

.. $id

If this information is not given by user at the time of creation
of 'ASpace', an empty list will be supplied.

          
The slot 'trip' contains information on trips made by the survey
respondent. It a data frame object with the following variables:

.. $tr_id: Trip ID (integer)\cr
.. $tr_seq: Sequence ID of the trip (integer)\cr
.. $purpose: Purpose of the trip (factor)\cr
.. $mode: Mode of the trip (factor)\cr
.. $o_type: Type of the origin (factor)\cr
.. $o_time: Time departed from the origin (integer, or POSIXct)\cr
.. $o_zone: Area code for the origin (character)\cr
.. $d_type: Type of the destination (factor)\cr
.. $d_time: Time arrived at the destination (integer, or POSIXct)\cr
.. $d_zone: Area code for the destination (character)\cr
}
\section{Slots}{

\describe{
\item{\code{info}}{an object of list. representing the demographic characteristic of individual.}

\item{\code{trip}}{an object of data.frame containing a trip data made by the person 
described in the info slot.}
}}

\examples{
#example of creating ASpace class
#creates sample info data
testinfo <- list()

#creates sample trip data
testtrip <- data.frame(tr_id = 1:10, tr_seq = rep(NA, 10), purpose = rep(NA, 10),
                      mode = rep(NA, 10), o_type = rep(NA, 10), 
                    o_time = c(200,300,400,500,600,700,800,900,1000,1100),
                    o_zone = c(1,2,3,4,5,6,5,4,3,2),
                    d_type = rep(NA, 10),
                    d_time = c(300,400,500,600,700,800,900,1000,1100,1200),
                    d_zone = c(2,3,4,5,6,5,4,3,2,1))
                    
#constructs an object of class 'ASpace'
testaspace <- new("ASpace", info = list(id = 1), trip = testtrip)


}
\seealso{
\linkS4class{ASpaces}
}
\author{
Seong-Yun Hong (syhong@khu.ac.kr)
}
