% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASpaces-methods.R
\docType{methods}
\name{[[,ASpaces,ANY,ANY-method}
\alias{[[,ASpaces,ANY,ANY-method}
\title{generic method for class ASpaces}
\usage{
\S4method{[[}{ASpaces,ANY,ANY}(x, i)
}
\description{
generic method for class ASpaces
}
