% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{construct}
\alias{construct}
\title{construct a slice object from an expression matrix}
\usage{
construct(exprmatrix, cellidentity = NULL, projname = NULL)
}
\arguments{
\item{exprmatrix}{A numeric matrix of expression values; rows are genes, columns are cells; row names are official NCBI gene symbols, column names are cell ids/names}

\item{cellidentity}{A vector encoding the original identity of cells (e.g., time information or cell types); if not set, the default is a vector of "cell"}

\item{projname}{A string describing the analysis; if not set, the default value is "SLICE-" with timestamp information}
}
\value{
the constructed and initialized slice object
}
\description{
construct and initialize a slice object from an expression matrix
}
