% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{construct_from_seurat}
\alias{construct_from_seurat}
\title{construct a slice object from a Seurat object}
\usage{
construct_from_seurat(
  obj,
  assay = "RNA",
  cellidentity = NULL,
  projname = NULL,
  scale = FALSE
)
}
\arguments{
\item{obj}{A Seurat object containing sample data;}

\item{assay}{The assay to extract expression matrix. Default is RNA}

\item{cellidentity}{Either character indicating which meta data column to use or a character assay in the length of cell count ob the object}

\item{projname}{A string describing the analysis; if not set, the default value is "SLICE-" with timestamp information}
}
\value{
the constructed and initialized slice object
}
\description{
construct and initialize a slice object from a Seurat object
}
