% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{.locate}
\alias{.locate}
\title{Find Individual's Location by Time}
\usage{
.locate(trip, id, at, na.rm, silent)
}
\arguments{
\item{trip}{an object of data.frame containg person's continuous trip.}

\item{id}{a vector of length 1, indicating person's ID}

\item{at}{a numeric vector of length 1, specifying the time at which 
the location of an individual needs to be extracted}

\item{silent}{logical. If TRUE, if any trip records have NAs in the time
fields, a warning message will be provided.}
}
\value{
a single row data.frame containing correspond with 'at'
}
\description{
Find where an individual is located at a specific time during continuous trip
}
\examples{
#creates sample trip data
testtrip <- data.frame(tr_id = 1:10, tr_seq = rep(NA, 10), purpose = rep(NA, 10),
                      mode = rep(NA, 10), o_type = rep(NA, 10), 
                    o_time = c(200,300,400,500,600,700,800,900,1000,1100),
                    o_zone = c(1,2,3,4,5,6,5,4,3,2),
                    d_type = rep(NA, 10),
                    d_time = c(300,400,500,600,700,800,900,1000,1100,1200),
                    d_zone = c(2,3,4,5,6,5,4,3,2,1))
                    
#Find where at 800 via .locate
.locate(testtrip, 1, 600, na.rm = TRUE, silent = FALSE)
}
\seealso{
\code{\link{slice}}
}
\author{
Seong-Yun Hong (syhong@khu.ac.kr)
}
