% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FB}
\alias{FB}
\alias{FB}
\title{FB}
\format{An object of class \code{ExpressionSet} with 35376 rows and 79 columns.}
\usage{
FB

FB
}
\description{
TODO: add detailed documentation on this code object

TODO: add detailed documentation on this code object
}
\keyword{datasets}
