% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FB.sig}
\alias{FB.sig}
\title{FB.sig}
\format{An object of class \code{character} of length 1016.}
\usage{
FB.sig
}
\description{
TODO: add detailed documentation on this code object
}
\keyword{datasets}
