% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{getDiffGenes}
\alias{getDiffGenes}
\title{Detecting lineage dependent differentially expressed genes}
\usage{
getDiffGenes(profiles, diff.thresh = 0.1, log.base = 2, thresh.exp = 1,
  thresh.cell = 2, thresh.var = 0.5)
}
\arguments{
\item{profiles}{The lineage dependent gene expression profiles}

\item{diff.thresh}{The FDR threshold for significance}

\item{thresh.exp}{The expression threshold}

\item{thresh.cell}{The cell number threshold}

\item{thresh.var}{The variance threshold}
}
\value{
lineage dependent differentially expressed genes and their associated information, e.g., variance, pvalue, fdr, and M1 predicted expression
}
\description{
given the lineage dependent expression profiles extracted by the getProfiles function, this function first smoothes the expression profile of each gene using cubic regression splines,
fits to two models (M1: changes in the lineage; M0: no change),
and then uses likelihood ratio test to compare the goodness of fit of the two models and assigns a p-value to each gene
}
\details{
Genes meet the following criteria will be selected for likelihood ratio test: expression >=thresh.exp in at least thresh.cell cells and variance>=thresh.var
Genes with FDR<diff.thresh will be considered significant
}
