% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{getEntropy}
\alias{getEntropy}
\title{calculate scEntropies of individual cells}
\usage{
getEntropy(
  object,
  clusters = NULL,
  km = NULL,
  calculation = "bootstrap",
  exp.cutoff = 1,
  B.size = 1000,
  B.num = 1,
  clustering.k = 0,
  random.seed = 201602,
  ...
)
}
\arguments{
\item{object}{Slice object}

\item{km}{A symmetric matrix encoding the functional similarity of genes; the row names and column names must be official NCBI gene symbols.}

\item{calculation}{Choose the "bootstrap" or "deterministic" calculation of scEntropy}

\item{exp.cutoff}{The expression threshold of a gene to be considered expressed}

\item{B.size}{In bootstrap calculation, specifies the size of bootstrap sample; in deterministic calculation, specific the number of top expressed genes from each cell}

\item{B.num}{The number of bootstrap samples}

\item{clustering.k}{The number of functional clusters}

\item{random.seed}{The seed for randomness}

\item{cluster}{A data frame containing the functional clustering of genes; contains two columns "GENE" - official NCBI gene symbols and "CLUSTER" - cluster memberships.}
}
\value{
an slice object with calculated entropies in the "entropy" and "entropies" slots
}
\description{
performing bootstrap or deterministic calculation of scEntropy
}
