% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{getPatterns}
\alias{getPatterns}
\title{Discover lineage dependent temporal patterns}
\usage{
getPatterns(expr, do.zscore = T, c.method = "pam", k = NULL,
  d.method = function(x) as.dist((1 - cor(t(x)))/2), do.plot = T,
  plot.filename = NULL)
}
\arguments{
\item{expr}{The M1 predicted expression profiles of differentially expressed genes in a transition}

\item{do.zscore}{If TRUE, perform gene based zscore transformation prior to clustering}

\item{c.method}{The algorithm to find pattern clusters, currently support "pam" and "kmeans"}

\item{k}{The number of patterns to be discovered; if NULL, gap statistic (cluster::clusGap) will be used to determine an optimal k; parameters of cluster::clusGap are set as follows: k.max=10, B=50}

\item{d.method}{The distance method}

\item{do.plot}{Whether to plot discovered patterns}

\item{plot.filename}{The file name of the plot; if NULL, plot to the screen}
}
\value{
An object of class "pam" or "kmeans" representing the clustering results.
}
\description{
Given the M1 predicted expression profiles of differentially expressed genes in a transition, 
the function uses PAM clustering to divide differentially expressed genes into clusters (patterns)
}
