% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{getRDS}
\alias{getRDS}
\title{reduce expression space}
\usage{
getRDS(object, genes.use = NULL, method = "pca", num_dim = 2,
  log.base = 2, do.center = TRUE, do.scale = FALSE, use.cor = T,
  min.var = 0.5, min.cells = 3)
}
\arguments{
\item{object}{slice object}

\item{genes.use}{The set of genes whose expression will be used for dimension reduction; if NULL, set to all genes}

\item{method}{The method for dimension reduction; currently, only support "pca"; more options will be integrated soon}

\item{num_dim}{The number of dimensions to be extracted}

\item{log.base}{The base of logarithm}

\item{do.center}{If TRUE, substract row-mean from each value}

\item{do.scale}{If TRUE, divide by row standard deviation}

\item{use.cor}{If TRUE, use Pearson's correlation matrix as input to prcomp function}

\item{min.var}{The minimum expression variance for a gene to be included in dimension reduction}

\item{min.cells}{The minimum number of expressed cells for a gene to be included in dimension reduction}
}
\value{
updated slice object with reduced dimensions in the "rds" slot, the first two components are also used to set the "vds" slot for visualization.
}
\description{
perform dimesion reduction using PCA using the expression profiles of genes satisfied the following criteria:
   included in the "gene.use",
   had log2 expression variance greater than min.var, and
   expressed (>1) in at least "min.cells" number of cells.
}
