% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{getTrajectories}
\alias{getTrajectories}
\title{Reconstructing cell trajectories following a specific lineage (start -> end) in the inferred lineage model}
\usage{
getTrajectories(
  object,
  method = "sp",
  start,
  end,
  network = "mst",
  NN.threshold = 0.7,
  do.trim = F,
  do.stepwise = F,
  ...
)
}
\arguments{
\item{object}{Slice object}

\item{method}{Select either "sp" (shortest-path based) method or "pc" (principal-curve based) method to reconstruct cell trajectories following a lineage in the infered lineage model}

\item{start}{The numeric ID of the stable state from which the cell transition starts}

\item{end}{The numeric ID of the stable state at which the cell transition ends}

\item{network}{Used in the shortest path based method; "mst" means the shortest path will be explored on the minimum spanning tree of cells, while "csn" means on the cell similarity network (i.e., mst + local wiring)}

\item{NN.threshold}{Used in the shortest path based method, specifies the distance threshold for determining neighbors of each cell in the path. Default is 0.7 quantile of all edge weights in the network}

\item{do.trim}{Used in the principal curve based method. When set to TRUE, only the curve between the start and end stable states will be returned; otherwise, the whole curve will be returned.}
}
\value{
updated slice object with reconstructed cell trajectories in the "transitions" slot.
}
\description{
This function implements both shortest-path based and principal-curve based cell transitional path reconstruction.
}
\details{
Prerequisite: the "model" slot of the input slice object must be set
}
