% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{isPD}
\alias{isPD}
\title{Check if a matrix is PD}
\usage{
isPD(mat)
}
\arguments{
\item{mat}{A matrix. The input which is potentially non-PD.}
}
\value{
TRUE/FALSE whether the matrix is PD.
}
\description{
Use Cholesky decomposition to determine if a matrix is PD (faster than full eigendecomp).
}
\keyword{internal}
