% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{km}
\alias{km}
\title{km}
\format{An object of class \code{matrix} with 14280 rows and 14280 columns.}
\usage{
km
}
\description{
TODO: add detailed documentation on this code object
}
\keyword{datasets}
