% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{km_version}
\alias{km_version}
\title{km_version}
\format{An object of class \code{character} of length 1.}
\usage{
km_version
}
\description{
TODO: add detailed documentation on this code object
}
\keyword{datasets}
