% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{logL}
\alias{logL}
\title{Log likelihood function}
\usage{
logL(Sigma, invSigmahat)
}
\arguments{
\item{Sigma}{A matrix. The masked latent matrix.}

\item{invSigmahat}{A matrix. The estimated precision matrix.}
}
\value{
The log likelihood.
}
\description{
This function is used to calculate the log likelihood.
}
\keyword{internal}
