% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{logseq}
\alias{logseq}
\title{Generate a log sequence.}
\usage{
logseq(beg, end, len)
}
\arguments{
\item{beg}{A numeric. The beginning of the sequence}

\item{end}{A numeric. The end of the sequence}

\item{len}{A numeric. The number of values to return.}
}
\value{
A vector of logarithmically spaced values.
}
\description{
Returns logarthmically spaced sequence of values.
}
\keyword{internal}
