% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{makePD}
\alias{makePD}
\title{Make a matrix positive definite (PD)}
\usage{
makePD(mat)
}
\arguments{
\item{mat}{A matrix. The input which is potentially non-PD.}
}
\value{
The PD version of the matrix.
}
\description{
This function adds a small value to diagonal to force a matrix to be PD.
}
\keyword{internal}
