% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\docType{methods}
\name{plotEntropies}
\alias{plotEntropies}
\alias{plotEntropies,slice-method}
\title{Visualize and compare entropies}
\usage{
plotEntropies(object)

\S4method{plotEntropies}{slice}(object)
}
\arguments{
\item{object}{Slice object containing the entropies in object@entropies}
}
\description{
Visualize entropies using boxplot and compare using regression
}
