% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{plotEntropies}
\alias{plotEntropies}
\title{Visualize and compare entropies}
\usage{
plotEntropies(object, ...)
}
\arguments{
\item{object}{Slice object containing the entropies in object@entropies}
}
\description{
Visualize entropies using boxplot and compare using regression
}
