% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{sim_slice_data}
\alias{sim_slice_data}
\title{Simulate single-layer covariance data.}
\usage{
sim_slice_data(r = 2, p = 50, n = 500, seed = NULL)
}
\arguments{
\item{r}{A numeric. The rank of the latent component.}

\item{p}{A numeric. The number of variables.}

\item{n}{A numeric. The number of observations.}

\item{seed}{A numeric. Optional random seed for reproducibility.}
}
\value{
A list containing:
\item{L_star}{The true latent component matrix.}
\item{Sigma_star}{The true covariance matrix.}
\item{X}{The generated observation matrix.}
\item{Sigma}{The sample covariance matrix computed from \code{X}.}
\item{S_star}{The true sparse precision component matrix.}
}
\description{
Generates sparse and low-rank components along with the corresponding
covariance matrix and sampled observations for a single-layer graphical
model.
}
