% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SLICE-Package}
\alias{SLICE-Package}
\alias{SLICE-Package-package}
\title{Determining Cell Differentiation and Lineage based on Single Cell Entropy}
\description{
Determining Cell Differentiation and Lineage based on Single Cell Entropy
}
