% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slicedata}
\alias{slicedata}
\title{Syhthetic Patterns of Activity Space}
\format{
An object of clss ASpaces. The dataset contains 84 people's demographic attribute and
236 trip records.
\describe{
  \item{data}{include list object which contains info and trip data}
  \item{info}{14 different demographic characteristics, including individual ID}
  \item{trip}{10 essential information which is needed to configure ASpace class}
  \item{sp}{No sp data is given since it is synthetic travel information}
  ...
}
}
\usage{
data(slicedata)
}
\description{
A dataset containing the demographic attributes and trip record of almost
100 people
}
\examples{
#load data
data(slicedata)

#validate whether data has normal condition
validateASpaces(slicedata)

#check data's basic information
summary(slicedata)
show(slicedata)
}
\keyword{datasets}
