% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASpaces-methods.R
\name{[[<-,ASpaces,ANY,ANY-method}
\alias{[[<-,ASpaces,ANY,ANY-method}
\title{generic method for class ASpaces}
\usage{
\S4method{[[}{ASpaces,ANY,ANY}(x, i) <- value
}
\description{
generic method for class ASpaces
}
