% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset}
\alias{subset}
\title{Subsets the Input object of Class ASpaces}
\usage{
subset(x, condition, vars, all = TRUE, showProgress = TRUE)
}
\arguments{
\item{x}{an object of class \code{ASpaces}}

\item{condition}{a list that indicate characteristics.}

\item{vars}{an optional character vector. See details.}

\item{all}{logical. See details.}

\item{showProgress}{logical. If TRUE, a progress bar appears on the R console 
while iterating.}
}
\value{
An object of class ASpaces
}
\description{
Extract \code{ASpace} objects with characteristics that meet given conditions from \code{ASpaces}
}
\details{
\code{subset} function is a function that refines a given \code{ASpaces} 
class object.

\code{ASpace} classes contain the characteristics 
of the activity space subject in the info slot, and \code{subset} function 
extracts only \code{ASpaces} that meet the given conditions based on the 
characteristics of#' this info slot and returns them to \code{ASpaces}.

'condition' is entered in the form of a \code{list}, and the \code{ASpace} 
of the corresponding condition is extracted by entering each desired 
condition in the \code{list} (i.e., \code{list(var1 = x, var2 = y)})

'vars' are used to refine the information in the resulting \code{ASpace}'s 
\code{info} slot. When a specific character vector is input to 'vars', the 
\code{info} slot of \code{ASpace} resulting from the function has only the 
input variables. If 'vars' is not input, the result is same with the original 
\code{info} slot.

'all' determines whether to extract only the results that satisfy all of them 
when various conditions are given in the 'condition', or to extract the 
results that apply at least one. The default value is set to extract only 
when all conditions are satisfied as true.
}
\examples{

# load data
data(slicedata)

# subsetting ASpaces
subsetasp1 <- subset(slicedata, condition = list(sex = 1))
subsetasp2 <- subset(slicedata, condition = list(sex = 1, hhsize = 4))
subsetasp3 <- subset(slicedata, condition = list(sex = 1, hhsize = 4), 
                     all = FALSE)
                     
}
\author{
Changlock Choi (hihi7100@khu.ac.kr), Seong-Yun Hong (syhong@khu.ac.kr)
}
