% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASpaces-methods.R
\name{update,ASpaces-method}
\alias{update,ASpaces-method}
\title{generic method for class ASpaces}
\usage{
\S4method{update}{ASpaces}(object, data, attr, sp)
}
\description{
generic method for class ASpaces
}
