\name{Slice-package}
\alias{Slice-package}
\docType{package}
\title{
Slince Bayesian Inference Engine 
}
\description{
Bayesian Inference 

}
\details{
\tabular{ll}{
Package: \tab gpHist\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2019-09-29\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}


}
\author{

Maintainer: Dennis Becker <dbecker@leuphana.de>

}

\references{


}
\keyword{ package }

\seealso{

}
