\name{slice}
\alias{slice}
\title{Function Title}
\description{
This is a function description.
}
\usage{
slice(x, y, z)
}
\arguments{
  \item{x}{a numeric matrix or data frame with coordinates.}
  \item{y}{a type of the data.}
  \item{z}{a type of the data.}
  \item{\dots}{optional arguments to be passed to another function.}
}

\details{
The purpose of \code{slice} is this.
}

\value{
An object of \code{data.frame}.
}
\note{
An extra details come here.
}
\author{
Seong-Yun Hong
}
\seealso{
\code{\link{foo1}}, \code{\link{foo2}}, \code{foo3}
}

\examples{
# examples
x <- 1:10
slice(x)

\dontrun{
# examples that do not run
y <- runif(10)
slice(y)
}
}